# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FloraCare API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 4 */
  password: string;
}

export interface User {
  id: string;
  username: string;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface Scan {
  id: string;
  userId: string;
  imageUrl: string;
  plantName: string;
  diseaseName: string;
  confidence: number;
  isHealthy: boolean;
  description: string;
  symptoms?: string;
  treatment?: string;
  prevention?: string;
  scannedAt: string;
}

export interface DiseaseCount {
  diseaseName: string;
  count: number;
}

export interface ScanStats {
  totalScans: number;
  diseasedScans: number;
  healthyScans: number;
  recentScans: Scan[];
  topDiseases: DiseaseCount[];
}

export interface Disease {
  id: string;
  plantName: string;
  diseaseName: string;
  description: string;
  symptoms: string;
  treatment: string;
  prevention: string;
}

export interface ScanUpload {
  image: Blob;
}

export interface ErrorResponse {
  error: string;
}

export type ListDiseasesParams = {
  search?: string;
};