# lib/api-zod/src/generated/types/authInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FloraCare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AuthInput {
  /** @minLength 3 */
  username: string;
  /** @minLength 4 */
  password: string;
}