# lib/api-zod/src/generated/types/disease.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FloraCare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Disease {
  id: string;
  plantName: string;
  diseaseName: string;
  description: string;
  symptoms: string;
  treatment: string;
  prevention: string;
}