# lib/api-zod/src/generated/types/diseaseCount.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FloraCare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DiseaseCount {
  diseaseName: string;
  count: number;
}