# lib/api-zod/src/generated/types/scan.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FloraCare API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Scan {
  id: string;
  userId: string;
  imageUrl: string;
  plantName: string;
  diseaseName: string;
  confidence: number;
  isHealthy: boolean;
  description: string;
  symptoms?: string;
  treatment?: string;
  prevention?: string;
  scannedAt: string;
}