# lib/api-zod/src/generated/types/scanStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * FloraCare API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiseaseCount } from "./diseaseCount";
import type { Scan } from "./scan";

export interface ScanStats {
  totalScans: number;
  diseasedScans: number;
  healthyScans: number;
  recentScans: Scan[];
  topDiseases: DiseaseCount[];
}